refactor: Reorganize repo
