+++ base/CBaseModelInfo.hpp
