+++ base/CNavigation.hpp
