+++ network/CNetGamePlayer.hpp
