+++ network/CNetworkPlayerMgr.hpp
