+++ network/netObject.hpp
