+++ ped/CPedAppearanceDataNode.hpp
