+++ ped/CPedComponents.hpp
