+++ ped/CPedFactory.hpp
