+++ ped/CPedModelInfo.hpp
