+++ ped/CPedWeaponManager.hpp
