+++ player/CNonPhysicalPlayerData.hpp
