+++ player/CPlayerAppearanceDataNode.hpp
