+++ player/CPlayerInfo.hpp
