+++ rage/atSingleton.hpp
