+++ socialclub/FriendInfo.hpp
