+++ socialclub/FriendRegistry.hpp
