+++ socialclub/ScInfo.hpp
