+++ vehicle/CAutomobile.hpp
