+++ vehicle/CVehicleDrawHandler.hpp
