+++ vehicle/CVehicleModelInfo.hpp
