+++ weapon/CAmmoInfo.hpp
