+++ weapon/CItemInfo.hpp
