+++ weapon/CWeaponInfo.hpp
